+++ lib/templates/test/rspec/module/test/Gemfile.tt
@@ -4,6 +4,6 @@ source "https://rubygems.org"
 build_gemfile(
   "terraspace",
   "rspec-terraspace",
-  provider_gem_name,
+  plugin_gem_name,
 )
 %>
