+++ lib/terraspace_plugin_google/autoloader.rb
@@ -1,6 +1,6 @@
 require "zeitwerk"
 
-module TerraspaceProviderGoogle
+module TerraspacePluginGoogle
   class Autoloader
     class Inflector < Zeitwerk::Inflector
       def camelize(basename, _abspath)
@@ -15,7 +15,7 @@ module TerraspaceProviderGoogle
         loader.inflector = Inflector.new
         lib = File.expand_path("../", __dir__)
         loader.push_dir(lib)
-        loader.ignore("#{lib}/terraspace_provider_google.rb")
+        loader.ignore("#{lib}/terraspace_plugin_google.rb")
         loader.setup
       end
     end
