+++ lib/terraspace_plugin_google/version.rb
@@ -1,0 +1,3 @@
+module TerraspacePluginGoogle
+  VERSION = "0.0.0"
+end
