--- lib/terraspace_provider_google/version.rb
@@ -1,3 +1,0 @@
-module TerraspaceProviderGoogle
-  VERSION = "0.0.0"
-end
