+++ spec/terraspace_provider_spec.rb
@@ -1,5 +1,5 @@
-RSpec.describe TerraspaceProviderGoogle do
+RSpec.describe TerraspacePluginGoogle do
   it "has a version number" do
-    expect(TerraspaceProviderGoogle::VERSION).not_to be nil
+    expect(TerraspacePluginGoogle::VERSION).not_to be nil
   end
 end
